chore: combine TestingHelper and ComponentTestingHeler
